.github/actions/smart-ci/requirements.txt
PyGithub==1.59
pyyaml==6.0.1
jsonschema==4.19.1